requirements_test.txt
# Test dependencies for Kwikset Smart Locks integration
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-homeassistant-custom-component>=0.13.0
